create api fetch mixin
